+++ packages/terra-overlay/tests/wdio/loading-overlay-spec.js
@@ -1,23 +1,20 @@
-describe('Loading Overlay', () => {
+Terra.describeViewports('Loading Overlay', ['huge'], () => {
   describe('Loading Overlay Default', () => {
     before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/default-loading-overlay'));
 
-    Terra.it.isAccessible({ selector: '#terra-LoadingOverlay' });
-    Terra.it.matchesScreenshot({ selector: '#terra-LoadingOverlay' });
+    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
   });
 
   describe('Loading Overlay Container', () => {
     before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/container-loading-overlay'));
 
-    Terra.it.isAccessible({ selector: '#terra-LoadingOverlay' });
-    Terra.it.matchesScreenshot({ selector: '#terra-LoadingOverlay' });
+    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
   });
 
   describe('Loading Custom Message', () => {
     before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/custom-message-loading-overlay'));
 
-    Terra.it.isAccessible({ selector: '#terra-LoadingOverlay' });
-    Terra.it.matchesScreenshot({ selector: '#terra-LoadingOverlay' });
+    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
 
     it('Expect default to have full screen', () => {
       expect(browser.getAttribute('#terra-LoadingOverlay', 'class')).contains('fullscreen');
@@ -35,21 +32,18 @@ describe('Loading Overlay', () => {
   describe('Light Loading Theme', () => {
     before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/light-loading-overlay'));
 
-    Terra.it.isAccessible({ selector: '#terra-LoadingOverlay' });
-    Terra.it.matchesScreenshot({ selector: '#terra-LoadingOverlay' });
+    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
   });
 
   describe('Dark Loading Overlay Theme', () => {
     before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/dark-loading-overlay'));
 
-    Terra.it.isAccessible({ selector: '#terra-LoadingOverlay' });
-    Terra.it.matchesScreenshot({ selector: '#terra-LoadingOverlay' });
+    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
   });
 
   describe('Clear Loading Overlay Theme', () => {
     before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/clear-loading-overlay'));
 
-    Terra.it.isAccessible({ selector: '#terra-LoadingOverlay' });
-    Terra.it.matchesScreenshot({ selector: '#terra-LoadingOverlay' });
+    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
   });
 });