+++ packages/terra-image/CHANGELOG.md
@@ -3,8 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+3.12.0 - (July 16, 2019)
+------------------
 ### Added
 * Added documentation on required peerDependencies
+
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 
 3.11.0  - (July 11, 2019)
 ------------------