+++ packages/terra-tag/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.16.0 - (July 16, 2019)
+------------------
 ### Added
 * Added documentation on required peerDependencies