+++ packages/terra-toggle/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+3.15.0 - (July 16, 2019)
+------------------
 ### Added
 * Added documentation on required peerDependencies